Add CONFIGURE_DEPENDS to CMake GLOB lists
The addition of new source files will not ordinarily trigger them to be
added to the build due to how GLOBs are implemented in CMake.

CMake 3.12 added a new CONFIGURE_DEPENDS flag to enable CMake to
re-check GLOB results during build time and re-generate the build
scripts if necessary. The existing CMakeLists.txt already specifies a
minimum CMake version of 3.13 so this should always be available.