Fix unit test execution when building in shared linking mode
The unit test is built only when static linking, but the test
itself was still being invoked when running 'ctest' in shared
linking mode. This leads to the following test failure in 'ctest':

'Could not find executable vvenc_unit_test'.

This adds a fix so the unit test is not run in shared builds.