json: don't directly link to target
The issue with directly linking to nlohmann_json::nlohmann_json is that
it populates the `INTERFACE_LINK_LIBRARIES` property on the exported
vvenc library, which would cause downstream projects from failing to
find vvenc the vvencTargets*.cmake files do not run a
`find_package(nlohmann_json)` call.

However, in this case, there's not much reason to fix it in the export
files, since nlohmann_json is only used as a header library for internal
headers, and none of the public headers require it to be included. Thus,
we can work around this by only pulling the cflags of the target and
ignoring any link flags.

Signed-off-by: Christopher Degawa <ccom@randomderp.com>