Add unit tests for MCTF motionErrorLumaInt8
In #465 we added a Neon implementation of motionErrorLumaInt8, however
this was not covered by any of the existing unit tests. To try and
ensure that the behaviour of the x86 and Arm kernels continues to match,
add a unit test covering these kernels.

To enable the tests to call the optimized kernels in a way matching how
we previously did this for TCoeffOps, mark the MCTF member kernels as
public members of the class and pass a new enableOpt member to allow us
to compare the optimized and fallback implementations.

Also add -Wno-sign-compare when compiling unit tests with GCC to avoid
warnings from Common.h and Slice.h, since this flag is already present
when building the library.