vvenc_unit_test: Enable --seed as a command line argument
If the unit tests crash part way through it is impossible to recover the
seed that was used since it was never printed. To get around this, print
the seed at the start of the test before actually running anything.

To enable re-running tests with the seed that failed, extend the
existing unit test `main` program with command line argument parsing
based on the existing logic in `ParseArg.h`. For now this only supports
`--help` and `--seed=<value>`.