Add unit tests for InterPrediction xFpBiDirOptFlow
We have both x86 and Arm implementations of this kernel, so add some
simple unit tests to ensure that the behaviour matches between them.

This requires us to extract the C fallback implementation for
xFpBiDirOptFlow to its own function and wire it up as the default
function pointer. As part of this we also need to move xRightShiftMSB to
be a non-member function so it is callable from the new fallback
implementation.

Similar to other unit tests, we add a new `enableOpt` parameter to the
`InterPredInterpolation::init` function to allow us to construct both a
reference and optimized object to compare against in the test.