Set -ffp-contract=off to aid cross-platform bit-reproducability
Currently when running VVenC on Arm machines, the md5sum of the output
video differs from x86 even when specifying `--SIMD=SCALAR`. This is due
to the default `-ffp-contract` setting on GCC and LLVM enabling fused
multiply-add instructions for floating-point arithmetic, which causes
small bit differences due to the lack of intermediate rounding in these
instructions.

Setting `-ffp-contract=off` disables this behaviour, ensuring the md5sum
output for the tested video now matches for `--SIMD=SCALAR`. There are
some remaining differences when other SIMD settings are used on Arm
which will be resolved in a separate PR.

Change-Id: Id6f6eb08e6e3ed269281ed9c9e017159b760cae0