arm: Rename RdCostARM.h => RdCost_neon.cpp and simplify
This file is currently templated over `ARM_VEXT`, however in practice
this can only be set to `NEON`, and a future SVE implementation would
need other changes.

In preparation for future changes to this file, simplify it to match the
others in the `neon/` directory by moving the content to the `.cpp` file
and removing the unused `ARM_VEXT` templating.

There is no need for these functions to be members of the `RdCost`
class, so remove that, however this means that the x86 implementations
of these kernels need to become public members of the class until we can
remove the dependency on them from the Neon kernels.