Rewrite xGetSADwMask_neon to improve performance
The subtraction and absolute value can be replaced with a single
absolute difference instruction, and we can use a pair of accumulators
to avoid a bottleneck on the accumulating multiply-add instruction
latency.

When benchmarking with LLVM-20 on recent Arm Neoverse
micro-architectures, this reduces the time spent in xGetSADwMask_neon by
~19%.

Additionally, enable this kernel even when TARGET_SIMD_X86 is disabled,
since it does not rely on the x86 intrinsics code.