vvenc_unit_test: Add --filter to run a subset of unit tests
When developing locally it can be useful to only run a subset of the
unit tests for faster test turnaround.

This commit adds a new --filter parameter that can be used to specify
the test suite to run. To enable this we organise the existing suites
into an array of entries and iterate through them, checking against the
specified --filter (or run everything if --filter is not specified
giving "").