vvenc_unit_test: Add --testcase/-t to run a subset of unit tests
When developing locally it can be useful to only run a subset of the
unit tests for faster test turnaround.

This commit adds a new --testcase/-t parameter that can be used to
specify the test suite to run. To enable this we organise the existing
suites into an array of entries and iterate through them, checking
against the specified testcase (or run everything if --testcase/-t was
not specified giving "").