Arm: Add SVE implementations of InterpolationFilter::filterWxH_N8
Add SVE implementations and corresponding init function for the 8-tap
filterWxH_N8.

The 8-tap filter for filter4x4 is reduced to a 6-tap filter due to the
zero coefficients at the 0th and 7th indices.

These SVE implementations are around 26% faster than their Neon
counterparts when benchmarked on a Neoverse V2 using LLVM 20.