Optimize the Neon implementation of xGetSAD_NxN
Optimize and simplify the Neon implementation of xGetSAD_NxN.  Use a
generic function that can take a runtime-defined value for width so that
it can also be used in xGetHAD2SADs_neon.
Remove the logic around early exit since the equivalent scalar functions
(xGetSAD<width>) do not make use of it.

These optimizations also allow these functions to be used even when
TARGET_SIMD_X86 is not defined.

These changes make the getSAD functions 1.1x faster on average depending
on the block size, when measured on a Neoverse V2 machine with LLVM-21.